pyproject.toml
[project]
name = "tfmod"
version = "1.0.0"
description = "A tool for managing public Terraform modules"
authors = [
  {name = "Josh Holbrook", email = "josh.holbrook@gmail.com"}
]
readme = "README.md"
keywords = []
license = { text = "BSD-3" }
classifiers = [
    "Programming Language :: Python",
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries",
]
requires-python = ">=3.12"
dependencies = [
  "go-flag",
  "pygithub>=2.5.0",
  "pyyaml>=6.0.2",
  "rich>=13.9.4",
  "tf-registry",
  "toml",
]

[project.optional-dependencies]
dev = [
  "flake8",
  "flake8-black",
  "pytest",
  "black",
  "isort",
  "jupyter-console",
  "validate-pyproject[all]",
]

[project.scripts]
router = "tfmod:main"

[tool.uv]
package = true

[tool.isort]
profile = "appnexus"
known_application = "tfmod"

[tool.pyright]
include = ["tfmod", "tests"]

[tool.pytest]
addopts = "--verbose -s"
testpaths = [ "tests" ]

[tool.setuptools]
packages = [ "tfmod" ]
